Implement: handle error for database connect

+++ src/app/modules/chats/infra/ChatRepositoryImpl.ts
@@ -6,6 +6,11 @@ import { collection } from "@/app/modules/chats/infra/data-access/MongoDB";
 export class ChatRepositoryImpl implements ChatRepository{
 
     async save(chat: Chat, newMessage: any): Promise<ObjectId | null>{
+        if(!collection){
+            console.warn(`Database is not connected. Chat will not be save`);
+            return null;            
+        }
+
         const filter = { _id: chat._id };
         const update = { 
             $setOnInsert: { chat },
@@ -23,17 +28,26 @@ export class ChatRepositoryImpl implements ChatRepository{
     };
 
     async getAll(): Promise<Chat[] | null>{
+        if(!collection){
+            console.warn(`Database is not connected`);
+            return null;            
+        }
+
         try {
             const chats = collection.find();
             return await chats.toArray() as Chat[];
         } catch (error) {
             console.error(`Error occurred while getting all chats: ${error}`);            
             throw error;
-        }
-        
+        }        
     };  
 
     async get(id: ObjectId): Promise<Chat | null>{
+        if(!collection){
+            console.warn(`Database is not connected`);
+            return null;            
+        } 
+
         try {
             const chat = await collection.findOne({_id: id}); 
             return chat as Chat;
@@ -44,6 +58,11 @@ export class ChatRepositoryImpl implements ChatRepository{
     };     
    
     async delete(id: ObjectId): Promise<number>{
+        if(!collection){
+            console.warn(`Database is not connected`);
+            return 0;            
+        }
+
         try {
             const result = await collection.deleteOne({_id: id});
             return result.deletedCount;
